elf parser: eliminate the some duplicated lines in little endian and big endian processing
SectionFlags comes from bit fields.
To parse them createFields() does similar in
little endian and big endian processing; the difference is
only the order of parsing.

This commit puts the field parsing into a tuple of lambdas. The
lambdas are applied for parsing the fields.
So just reversing the tuple is enough for processing difference
endian. As the result, some duplicated lines are eliminated.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>