lzx: fix decompression of window refs to Intel jump fixups
Intel jump fixups should only be applied to the final output, not to the LZX
window. We were incorrectly copying fixed-up bytes to the output, which caused
rare failures in the decompression process.

Fixes #76, and includes testcases to ensure the correct decompression of
georgia.cab and verdan32.exe (run with `SLOW_TESTS=1`).